Final Project update week 4 day 3 midday

+++ Workspace2/FinalProject/src/main/java/com/revature/util/GetAccountNumber.java
@@ -9,7 +9,7 @@ public class GetAccountNumber {
 	private static Logger log=Logger.getLogger(GetAccountNumber.class);
 
 	
-	public static double getAccountNumber(String message) {
+	public static int getAccountNumber(String message) {
 		log.info(message);
 		int accountNumber;
 		try {
